move logic to AppFrame
